src/main/java/com/foodsafety/importfood/osint/commons/CrawlerService.java
package com.foodsafety.importfood.osint.commons;

import com.microsoft.playwright.Browser;
import com.microsoft.playwright.Page;
import com.microsoft.playwright.Playwright;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
public class CrawlerService {


    public Page getPage(String url) {
        Playwright playwright = Playwright.create();
        Browser browser = playwright.chromium().launch();
        Page page = browser.newPage();
        page.navigate(url);
        browser.close();
        playwright.close();
        return page;
    }



}
